The Contact constructor should not set id / lastUpdated attributes
The Contact constructor should not set id / lastUpdated attributes. Those should be
initially null and generate asynchronously by the call to ContactManager.save()
because this is when the implementation actually interacts with the database. The
specification actually already states that the save() method will set the id and
lastUpdated attributes on the contact that is asynchronously returned.

This issue was reported by Huang Deqing <deqing.huang@intel.com>.